Add OrderList widget for managing orderable lists
Created a new OrderList widget to enable draggable and droppable lists with user preferences. Removed duplicate OrderList implementation from gui_prefs.py and refactored to use the new widget. This enhances maintainability and modularity of the codebase.
